+++ packages/reown_walletkit/example/lib/models/chain_metadata.dart
@@ -9,7 +9,6 @@ enum ChainType {
   polkadot,
   bitcoin,
   tron,
-  sui,
   stacks,
 }